+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Private objects should return 404 by api [#3311](https://github.com/opendatateam/udata/pull/3311) [#3316](https://github.com/opendatateam/udata/pull/3316)
 - Allow returning full sub-objects (license, frequency, zones and granularity) for datasets APIv2 [#3310](https://github.com/opendatateam/udata/pull/3310)
 - Add `featured` to dataset default mask [#3313](https://github.com/opendatateam/udata/pull/3313)
+- Add endpoint to get all schemas for a dataset without fetching resources [#3314](https://github.com/opendatateam/udata/pull/3314)
 
 ## 10.3.2 (2025-05-06)